Add empty `PeopleDirectoryMap` component
