+++ packages/lesswrong/lib/components.ts
@@ -1138,6 +1138,8 @@ importComponent("PeopleDirectorySearchableFilter", () => require('../components/
 importComponent("PeopleDirectorySelectOption", () => require('../components/peopleDirectory/PeopleDirectorySelectOption'));
 importComponent("PeopleDirectoryClearAll", () => require('../components/peopleDirectory/PeopleDirectoryClearAll'));
 importComponent("PeopleDirectoryResults", () => require('../components/peopleDirectory/PeopleDirectoryResults'));
+importComponent("PeopleDirectoryResultsMap", () => require('../components/peopleDirectory/PeopleDirectoryResultsMap'));
+importComponent("PeopleDirectoryResultsList", () => require('../components/peopleDirectory/PeopleDirectoryResultsList'));
 importComponent("PeopleDirectoryNoResults", () => require('../components/peopleDirectory/PeopleDirectoryNoResults'));
 importComponent("PeopleDirectoryResultRow", () => require('../components/peopleDirectory/PeopleDirectoryResultRow'));
 importComponent("PeopleDirectoryHeading", () => require('../components/peopleDirectory/PeopleDirectoryHeading'));
